set id to Long

+++ src/main/java/com/example/sintomedic/repositorios/UsuariosRepositorio.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuariosRepositorio extends JpaRepository<Usuario, Integer> {
+public interface UsuariosRepositorio extends JpaRepository<Usuario, Long> {
 
 }
